Merged Ruby code generation into main protoc binary.
There's no reason to build a separate protoc-gen-ruby plugin binary if
this language support is integrated into the protobuf distribution
proper. IMHO we should take this approach for the language generation
and ensure that the Ruby extension in ruby/ only installs a Ruby
extension .so, nothing else. This split is how Python works today too.